Revert change to forked CGLIB code

+++ spring-core/src/main/java/org/springframework/cglib/proxy/Enhancer.java
@@ -1258,7 +1258,7 @@ public class Enhancer extends AbstractClassGenerator {
 			}
 
 			public int getIndex(MethodInfo method) {
-				return (Integer) indexes.get(method);
+				return ((Integer) indexes.get(method)).intValue();
 			}
 
 			public void emitCallback(CodeEmitter e, int index) {
